refactor: share 폴더 제거, uikit 폴더 추가
