+++ src/App.tsx
@@ -1,7 +1,7 @@
 import { ThemeProvider } from 'styled-components';
 import { InitStyle, theme, GlobalFont } from '@/globalStyle';
 import { Outlet } from 'react-router-dom';
-import { Header } from '@/component/@share';
+import { Header } from '@/component/uikit';
 import { Footer } from '@/component/Footer';
 // import { ScrollToTop } from '@/util';
 import { useScrollToTop } from '@/hooks';
