+++ src/component/Footer/organism/Footer.tsx
@@ -2,13 +2,13 @@ import { useState } from 'react';
 import styled from 'styled-components';
 import { Mail, Privacy, CopyRight } from '@/component/Footer/atoms';
 import { OrganizationInfo } from '@/component/Footer/data';
-import { Toast } from '@/component/@share';
+import { Toast } from '@/component/uikit';
 import { devices } from '@/globalStyle';
 import { useWindowResize } from '@/hooks';
 import { ToastStatus } from '@/types/Toast';
 import { useRecoilState } from 'recoil';
 import { currentToastType, toastState } from '@/store';
-import { SquareButton } from '@/component/@share/atom/SquareButton';
+import { SquareButton } from '@/component/uikit/atom/SquareButton';
 import { useGaApi } from '@/hooks/useGA';
 import { useLocation, useSearchParams } from 'react-router-dom';
