+++ src/component/timeBlock/atoms/TimeBlock.tsx
@@ -1,7 +1,7 @@
 import { isMouseDownState } from '@/store/atoms';
 import { useRecoilState } from 'recoil';
 import styled from 'styled-components';
-import { Toast } from '@/component/@share';
+import { Toast } from '@/component/uikit';
 import { toastState, currentToastType } from '@/store/atoms';
 
 interface TimeBlockProps {
