+++ src/component/uikit/atom/LoadingIcon.tsx
