+++ src/component/uikit/molecules/TitleCheck.tsx
@@ -1,6 +1,6 @@
 import { PropsWithChildren } from 'react';
 import styled from 'styled-components';
-import { CheckBox } from '@/component/@share';
+import { CheckBox } from '@/component/uikit';
 import { theme } from '@/globalStyle';
 
 interface TitleCheckProps {
