+++ src/component/uikit/organisms/Header.tsx
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 import { useNavigate, useSearchParams } from 'react-router-dom';
-import { Logo, Helper } from '@/component/@share/atom';
+import { Logo, Helper } from '@/component/uikit/atom';
 import { useEffect, useState } from 'react';
 import { HelperModal } from './HelperModal';
 import { useLocation } from 'react-router-dom';
