+++ src/component/uikit/template/LoginMasterTemplate.tsx
@@ -1,8 +1,8 @@
 import styled from 'styled-components';
 import { useNavigate } from 'react-router-dom';
-import { TitleBox } from '@/component/@share/molecules';
-import { Input } from '@/component/@share/atom';
-import { ButtonLarge } from '@/component/@share/atom';
+import { TitleBox } from '@/component/uikit/molecules';
+import { Input } from '@/component/uikit/atom';
+import { ButtonLarge } from '@/component/uikit/atom';
 import { useState, useEffect } from 'react';
 import { useRecoilState, useResetRecoilState } from 'recoil';
 import { signUpNickname, signUpPassword } from '@/store/atoms/Login';
